# TweetHere-API-GATEWAY/cmd/docs/swagger.yaml
definitions:
  models.AdminLogin:
    properties:
      email:
        type: string
      password:
        maxLength: 20
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  models.ChangePassword:
    properties:
      new_password:
        type: string
      old_password:
        type: string
      re_password:
        type: string
    type: object
...
  models.UserLogin:
    properties:
      email:
        type: string
      password:
        maxLength: 20
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  models.UserOTPLogin:
    properties:
      email:
        type: string
    type: object
...
  models.UserSignup:
    properties:
      date_of_birth:
        type: string
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Tweet Here is a social Media Platform
  title: Go + Gin SocialMedia TweetHereeee
  version: 1.0.0
paths:
  /admin/login:
    post:
      consumes:
      - application/json
      description: Login handler for Zsoxial admins
      parameters:
      - description: Admin login details
        in: body
        name: adminDetails
        required: true
        schema:
          $ref: '#/definitions/models.AdminLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin Login
      tags:
      - Admin
  /admins/block:
    patch:
      consumes:
      - application/json
      description: Blocks a user by ID
      parameters:
      - description: User ID to block
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Block User
      tags:
      - Admin
  /admins/unblock:
    patch:
      consumes:
      - application/json
      description: Unblocks a user by ID
      parameters:
      - description: User ID to unblock
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Unblock User
      tags:
      - Admin
  /admins/userdetails:
    get:
      consumes:
      - application/json
      description: Retrieves user details for admin management
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - bearer: []
      summary: Get User Details
      tags:
      - Admin
  /user/login:
    post:
      consumes:
      - application/json
      description: Login handler for users
      parameters:
      - description: User login details
        in: body
        name: userDetails
        required: true
        schema:
          $ref: '#/definitions/models.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Login
      tags:
      - User
  /user/otplogin:
    post:
      consumes:
      - application/json
      description: Generates OTP for user login
      parameters:
      - description: User OTP login details
        in: body
        name: userDetails
        required: true
        schema:
          $ref: '#/definitions/models.UserOTPLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User OTP Login
      tags:
      - User
  /user/signup:
    post:
      consumes:
      - application/json
      description: Sign up handler for new users
      parameters:
      - description: User sign up details
        in: body
        name: userDetails
        required: true
        schema:
          $ref: '#/definitions/models.UserSignup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Sign Up
      tags:
      - User
  /users/acceptfollowreq:
    post:
      consumes:
      - application/json
      description: Accepts a follow request from another user
      parameters:
      - description: Logged-in User ID
        in: header
        name: id
        required: true
        type: integer
      - description: User ID to accept follow request from
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Accept Follow Request
      tags:
      - User
...
  /users/changepassword:
    patch:
      consumes:
      - application/json
      description: Changes the password of the logged-in user
      parameters:
      - description: User ID
        in: header
        name: id
        required: true
        type: integer
      - description: New password details
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/models.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Change Password
      tags:
      - User
...
  /users/followers:
    get:
      consumes:
      - application/json
      description: Retrieves the followers of the logged-in user
      parameters:
      - description: Logged-in User ID
        in: header
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Followers
      tags:
      - User
  /users/followings:
    get:
      consumes:
      - application/json
      description: Retrieves the users that the logged-in user is following
      parameters:
      - description: Logged-in User ID
        in: header
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Followings
      tags:
      - User
  /users/followreq:
    post:
      consumes:
      - application/json
      description: Sends a follow request to another user
      parameters:
      - description: Logged-in User ID
        in: header
        name: id
        required: true
        type: integer
      - description: User ID to send follow request to
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Send Follow Request
      tags:
      - User
...
  /users/getyoudetails:
    get:
      consumes:
      - application/json
      description: Retrieves details of the logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get your Details
      tags:
      - User
...
securityDefinitions:
  bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"